refactored scss starting progress bar
