+++ src/javascripts/components/progressBar/progressBar.js
@@ -1,0 +1,13 @@
+import './progressBar.scss';
+// import eat from '../quadrants/eat';
+import play from '../quadrants/play';
+// import fight from '../quadrants/fight';
+// import sleep from '../quadrants/sleep';
+
+// const health = eat.full + play.fun + fight.strength + sleep.energy;
+
+const printProgressBar = () => {
+  console.log(play.fun);
+};
+
+export default { printProgressBar };
