+++ src/javascripts/components/progressBar/progressBar.scss
@@ -1,0 +1,28 @@
+@import '../../../styles/colors';
+@import '../../../styles/variables';
+
+#fullScore, #strengthScore{
+    margin-top: 7px;
+}
+
+#funScore, #energyScore{
+    display: flex;
+    justify-content: flex-end;
+    margin-top: 7px;
+}
+
+.progressBar{
+    background-color: $lightblue;
+    border: 1px solid dimgrey;
+    border-radius: 20px;
+    margin-right: 5px;
+    width: 40%;
+}
+
+#eatBarPercent, #fightBarPercent{
+    @include progress-fill(100%);
+}
+
+#playBarPercent, #sleepBarPercent{
+    @include progress-fill(50%);
+}
