+++ src/javascripts/main.js
@@ -1,10 +1,11 @@
 import '@fortawesome/fontawesome-free/js/all';
 import '../styles/main.scss';
-import eat from './components/eat';
-import fight from './components/fight';
+import eat from './components/quadrants/eat';
+import fight from './components/quadrants/fight';
 import pet from './components/pet/pet';
-import play from './components/play';
-import sleep from './components/sleep';
+import play from './components/quadrants/play';
+import sleep from './components/quadrants/sleep';
+import health from './components/progressBar/progressBar';
 
 const init = () => {
   eat.printEatBtns();
@@ -12,6 +13,7 @@ const init = () => {
   pet.printPet();
   play.printPlayBtns();
   sleep.printSleepBtns();
+  health.printProgressBar();
 };
 
 init();
