+++ src/styles/_colors.scss
@@ -1,0 +1,7 @@
+$blue: rgb(0, 81, 128);
+$lightblue: rgba(0, 81, 128, 0.616);
+$category: #77918365;
+$eat: #c5dbcf;
+$play: #b8d4c5;
+$fight: #acc9ba;
+$sleep: #a1bbae;
